+++ src/main/java/model/judgement/DealerResult.java
@@ -1,4 +1,4 @@
-package model;
+package model.judgement;
 
 public record DealerResult(
         int winCount,
