+++ src/test/java/model/JudgementTest.java
@@ -5,10 +5,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import model.card.Card;
+import model.judgement.GameStatus;
+import model.judgement.Judgement;
+import model.judgement.PlayerResult;
 import model.paticipant.Dealer;
 import model.paticipant.Player;
 import model.paticipant.Players;
-import model.service.Judgement;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
